+++ apps/deploy-web/src/pages/deploy-linux/index.tsx
@@ -1,7 +1,10 @@
 import { NewDeploymentContainer } from "@src/components/new-deployment/NewDeploymentContainer";
 import { withSdlBuilder } from "@src/context/SdlBuilderProvider/SdlBuilderProvider";
+import { getServerSideProps } from "../new-deployment";
 
 export default withSdlBuilder({
   componentsSet: "ssh",
   imageSource: "ssh-vms"
 })(NewDeploymentContainer);
+
+export { getServerSideProps };