+++ apps/provider-console/src/utils/urlUtils.ts
@@ -17,7 +17,9 @@ export const domainName = "https://provider.akash.network";
 export class UrlService {
   static home = () => "/";
   static deployments = () => "/deployments";
+  static attributes = () => "/attributes";
   static getStarted = () => "/get-started";
   static privacyPolicy = () => "/privacy-policy";
   static termsOfService = () => "/terms-of-service";
+  static actions = () => "/actions";
 }